feat: Implement media reference resolution for messages and update module structure
