+++ zihuan_service/src/storage/mod.rs
@@ -1,2 +1,3 @@
+pub(crate) mod media;
 pub(crate) mod qq_chat_history_store;
 pub(crate) mod qq_chat_session_store;
